Also trigger rescan if an attribute changes
This will help in the scenario where you copy music to the share but its
default permissions/ownership prevent Squeezebox Server from scanning
the files when the scan is triggered.

Now, when you've fixed the permissions on the files this re-triggers
another scan when the files will now be readable and they'll be found by
the server.